# src/docs/swagger.yaml
basePath: /
definitions:
  github_com_gothi_vouchrs_src_delivery_http_response.APIError:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: listing not found
        type: string
    type: object
  github_com_gothi_vouchrs_src_delivery_http_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.APIError'
      meta: {}
      success:
        example: true
        type: boolean
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequest:
    properties:
      alertedCount:
        type: integer
      brandID:
        type: string
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      maxPrice:
        type: number
      maxValue:
        type: number
      minValue:
        type: number
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequestStatus'
      userID:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequestStatus:
    enum:
    - active
    - fulfilled
    - expired
    - cancelled
    type: string
    x-enum-varnames:
    - BuyRequestStatusActive
    - BuyRequestStatusFulfilled
    - BuyRequestStatusExpired
    - BuyRequestStatusCancelled
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequest:
    properties:
      adminNotes:
        type: string
      brand:
        description: free-text brand name
        type: string
      createdAt:
        type: string
      desiredValue:
        type: number
      fulfilledAt:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestStatus'
      updatedAt:
        type: string
      urgency:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestUrgency'
      userID:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestStatus:
    enum:
    - PENDING_ADMIN_REVIEW
    - UNDER_REVIEW
    - FULFILLED
    - REJECTED
    - DEFERRED
    type: string
    x-enum-varnames:
    - CardRequestStatusPendingReview
    - CardRequestStatusUnderReview
    - CardRequestStatusFulfilled
    - CardRequestStatusRejected
    - CardRequestStatusDeferred
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestUrgency:
    enum:
    - flexible
    - within_24h
    - within_1_week
    type: string
    x-enum-varnames:
    - CardRequestUrgencyFlexible
    - CardRequestUrgency24h
    - CardRequestUrgency1Week
...
  src_delivery_http_handler.createBuyRequestBody:
    properties:
      brand_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      max_price:
        example: 900
        type: number
      max_value:
        example: 2000
        type: number
      min_value:
        example: 500
        type: number
    required:
    - brand_id
    - max_price
    - max_value
    - min_value
    type: object
  src_delivery_http_handler.createCardRequestBody:
    properties:
      brand:
        example: Amazon India
        type: string
      desired_value:
        example: 500
        type: number
      urgency:
        enum:
        - flexible
        - within_24h
        - within_1_week
        example: within_24h
        type: string
    required:
    - brand
    - desired_value
    - urgency
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cardswap.in
    name: CardSwap Support
  description: P2P gift card exchange marketplace. Card codes are AES-256 encrypted
    at rest and **never returned in API responses** — they are delivered to the buyer's
    registered email only.
  title: CardSwap India API
  version: "1.0"
paths:
...
  /api/v1/buy-requests:
    get:
      description: Returns all active buy requests created by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequest'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: List my buy requests
      tags:
      - requests
    post:
      consumes:
      - application/json
      description: Register interest in buying a specific brand + value range. CardSwap
        will notify you by email when a matching listing goes live.
      parameters:
      - description: Buy request criteria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.createBuyRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Create a buy request
      tags:
      - requests
  /api/v1/buy-requests/{id}:
    delete:
      description: Remove a buy request. Only the owner can delete it.
      parameters:
      - description: Buy request UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Not the owner
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a buy request
      tags:
      - requests
  /api/v1/card-requests:
    post:
      consumes:
      - application/json
      description: Ask CardSwap to source a specific gift card on your behalf. Admin
        reviews and fulfils these requests manually. Urgency affects prioritisation.
      parameters:
      - description: Card request details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.createCardRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Create a card request (admin queue)
      tags:
      - requests
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"